Fixed bug when a state having no message matching clauses wasn't recognized as a parse error.
Made a significant change in message matching. Previously, badly formed
messages coming from the environment or adversary had to be explicitly
matched using othermsg. But there's nothing sensible to do in response
other than to fail. So this is now the action that implicitly happens,
and there is no way to take another action.