Refactored message matching code to unify treatment of functionalities and simulators.
In the process: changed meaning of "*" in initial state of a simulator.

Now it applies relative to possible messages that could be received
(only ones from ideal functionality).  So saying "*" is legal. (Before
there was an error message that messages from adversary to real
functionality could not be intercepted in initial state. So one had to
say <IdealFun>.* to catch all over possible messages. Now "*" will
suffice.) There are, however, special cases in error messages for
simulators when a message path pattern is good for non-initial states,
but bad for the initial state.

To do this refactoring, added boolean saying whether in initial state
to state_context, and passed state_context to message matching code
that now needed it.