Code refactoring so that state machine checking in real functionalities, ideal functionalities and simulators uses completely common code. By including or not "simulator" in the state_context tags, and also setting the internal_ports argument appropriately, we can differentiate the behavior.
When checking the parties of a functionality, there is a first pass to
turn the party_defs into a map of party_def's, so we know their
names are distinct, and we can turn the names into internal ports. But
the rest of party checking is done in a single pass, and this is
when party_def_tyd's get created.

When checking state machines, there is a first pass to check unique
state names and top-level properties, and build a map. Then there
is a second phase to check states individually at a lower level,
using the map, as states are mutually recursive.