One of the choices of the refactoring of the previous commit led to less good error message than before. Before checking below the top-levels of states, we want to know that each party of a real functionality serves legal inter id paths, and that the totality of the serves clauses is distinct and covers all possibilities. (This is because the checking of message match clauses depends upon the serves clauses, and if the latter are messed up, this checking of message match clauses will confuse the user.)
So now: when checking the parties of a real functionality, we first
check the top-level of each party in turn, which involves checking the
serves clauses in isolation, and then checking the top-level of the
party's state machine (one initial state, uniqueness of state names,
distinctness of parameters, variables and internal ports), producing a
typed party. Then a check is made that the totality of the serves
clauses is good.  The result is a map of typed parties. Only then do
we work our way through the parties once more, checking their state
machines at the lower level.

When checking the state machine of an ideal functionality or a simulator,
we do the top-level checking of all the states, and then do the lower-level
checking of the states (as with the previous commit).