Reworking of lexer. More things are now handled: strings for names of mixfix operators; parenthesized operators (which can be used in prefix position) or simply as data; numeric literals for int and real; adding some more tokens we'll need when we add EasyCrypt types and expressions to the parser.
EasyCrypt's lexer allows different kinds of tokens to run up against
each other, and has a somewhat complex algorithm for splitting them
apart.  We've chosen a simpler approach where the user will have to
insert whitespace to keep some symbols separate. But our lexer now
returns a single token, as opposed to a list of tokens.

The above algorithm of EasyCrypt's lexer, plus how the tokens are used
in its parser, means that only certain symbols are are allowed as
infix operators. Because we will be generating EasyCrypt code, we
needed to be able to reproduce this behavior in our lexer - which is
now done. Lexing of infix operators now either gives an error message
saying illegal, or results in a single LOP1-LOP4 or ROP4.