Reworked SMC case study to make use of EPDPs for encoding and decoding, making use of EPDP combinators (instead of our old encode ..., and match decode).
For this to work, we need a match statement, like in the development
branch of EasyCrypt (1.0-preview). Note that a match only destructs
the constructor; if one wants to also destruct a tuple, one must
use something like

  match u with
  | Some u => {
      (pt1, pt2, k) <- tr;
      ...
    }
  | None   => { fail. }
  end

(We could add a match statement that combined these steps, but
to make it easier to translate into EasyCrypt, let's keep it
as in EasyCrypt.)

Note the Emacs List file
uc-dsl/examples/smc-case-study/.dir-locals.el, which is why running
EasyCrypt in Emacs on the examples of this directory is able to
require the .ec files in ec-dsl/prelude. When running from the shell
you need to use -I ../../prelude to get this effect.