It turns out when using -I options to EasyCrypt, each time you add another -I, it gives that directory *higher* precedence. (I thought it was the other way around!) So I've made it be the same with -I arguments to ucdsl.
But when doing searches (in ucUtil.ml) for files, it's more convenient
to have the beginning of the last have higher precedence. Consequently,
the include dirs list is now kept so that earlier is higher
precedence (new -Is are added at beginning). When we give the
list to EasyCrypt as its include path, we reverse it.

Like in EasyCrypt, the EasyCrypt theory directories have lowest
precedence, then the include dirs, then the current directory.