I misunderstood the precedence of the semantics of -I options to EasyCrypt: in fact, later -I's take precedence over earlier ones. So I've replicated that behavior in UC DSL. But it's still nice to keep the include dirs list in order from highest precedence to lowest. (We reverse this list when giving it to EasyCrypt as its include path.)
The directory of the argument to ucdsl is added at the beginning
of the include dirs list, as if it was supplied to a final -I option.

Consequently, the message of the previous commit was wrong. Here's the
story: the search for .ec and .uc files (apart from the initial
argument to ucdsl, which is interpreted without search) uses
the following precedences:

  (+) highest: the UC DSL prelude

  (+) medium high: the current directory

  (+) medium: the -I's

  (+) lowest, in case of .ec files: the EasyCrypt theories