Work on adding EasyCrypt types and expressions to parser.
There are some reduce/reduce and shift/reduce conflicts that
have to be investigated.

We're now using ecParsetree for various types, so switched to
using, e.g., psymbol instead of id, and qsymbol instead of qid.
This led to various changes about localization.

I added upper and lowercase identifiers, and made some choices of
which categories of UCDSL identifiers are which. Interfaces,
functionalities, simulators, party names, and state names are required
to be upper identifiers, but local variables are state parameters
are required to be lower identifiers.

Apart from fixing the parser conflicts, what's left is to add match
statements. At least some of the new version of the case study is
already parsing, so progress!