Moved the old UC DSL code for incompletely dealing with EasyCrypt types and expressions to directory "old". This directory can eventually be deleted.
Updated UcTypedSpec to use EcTypes.ty and EcSymbols.symbol.

Reconnected typechecker (UcTypecheck), but commented out most of
UcTypecheck, setting main function to raise exception. Thus, if we get
past the parser, we now get an exception.

The plan is to now to step-by-step update the typechecker, uncommenting as
we go.