Because EcTyping is used by a number of modules, and we wanted to include more of its values in its module interface, making a copy of it with names ucEcTyping.ml/mli wasn't practical.
Consequently, this commit reflects a change in the way modifications
of EasyCrypt modules/interfaces are handled.

In the src directory, there are now some files that begin with ec*:
ecTyping.ml/mli, ecScope.ml/mli, ecCommands.ml/mli. These are the UC
DSL versions of the corresponding EasyCrypt modules/interfaces. In
ECsrc, those files are (must be!) missing, but the originals have been
moved to ECsrc/modified-for-ucdsl - making it easy to do diff's with
the UC DSL versions.

This means that even though EcTyping is used by some of the modules
in ECsrc, they are getting the module/interface in the *src* directory
instead.

- - -

Also made a copy of ECsrc/ecUserMessages.ml/mli =>
UcEcUserMessages.ml/mli. In this case, we want to keep the original
module/interface, because when we do an ec_require, we let EasyCrypt
issue its error messages as usual. But in ECsrc/ecUserMessages.ml/mli
we can change the error messages as we see fit.

Modified ucTypecheck.ml to catch type error exceptions
EcTyping.TyError, and use UcEcUserMessages.TypingError.pp_tyerror in
conjunction with UC DSL's type_error to display error messages. So
we'll have localized type error messages, and we can tweak those
error messages in the future to suit our users.