Removed the instances of lsymbols in UcTypedSpec: the locations were only superficially needed - when traced to their sources, in all cases they were no longer needed.
Because UcSpec works with psymbol's, it is actually natural to
call the types of located symbols used in UcTypecheck by the
name psymbol. So removed the type lsymbol and its uses.

In check_defs, to facilitate earlier debugging, there was a leftover
wildcard pattern that stopped the checks on functionalties and
simulators from running.  I took this out and tried a few tests
checking top-level things in the case study, like making errors in the
argument to the real functionality being simulated. It all worked.