We are now using transexp to typecheck EasyCrypt expressions (see the function check_expr).
Everything is done, and seems to be working (it even caught one type error
in KeyExchange.uc!), except for the match instruction.

Our grammar for send-and-transition instructions was incomplete: instead
of an lident after the @, e.g.,

send KEDir.Pt1.ke_rsp2(k2 ^ q1)@pt1 and transition Final

it's better to have an arbitrary expression of type port. So I made that
change. It doesn't affect the case study, of course.

(This allows, e.g., a conditional expression (say, expr ? pt1 : pt2)
for determining the port. Of course, we could insist that a variable
of type port, pt, was initialized to either pt1 or pt2, and then use
@pt, but it's no harder to have this flexibility.)

On the other hand, this reminds me that we should try to do something
using static analysis: make sure that when the user says @exp, that
exp denotes an external port. E.g., we don't want one Party to be able
to send a message directly to another one (which they can now
do). Hopefully we can add more to the state_analysis type to track
this?