Refactored so that we no longer directly use EcParsetree or EcTyping. These modules are still used when a file is ec_required.
Instead, UcSpec is now self-contained, including the defintions of
types and expressions (so we can change these for the DSL independently
from EasyCrypt).

There is a new module UcTypecheckTypesExprs that has what we need from
EcTyping (we can change this in the future if we want to do more
analysis as part of typechecking).

UcEcUserMessages has been stripped down to what we need and
renamed to UcTypesExprsErrorMessages. (EasyCrypt itself will use
EcUserMessages, as usual.)

There is now a module interface for UcUtils.

Added copyright for BU, with the same software license as EasyCrypt.