UC DSL identifiers are now restricted not to start with "_", "uc_" or "UC_". The names of functionality interfaces, functionalities, simulators and parties may not conflict with the names of EasyCrypt theories in the current scope's environment. This restriction is to avoid conflicts when generating EasyCrypt code from UC DSL specs.
For a similar reason, UCCore.ec has now been split into a part
(UCBasicTypes.ec) used by UC DSL, and the rest - which will only be
used by EasyCrypt code. In smc-next, Forwarding.ec is now split into
Forwarding_Code.eca (which requires UCBasicTypes), which is what we
want to generate from examples/smc-case-study/Forwarding.uc, and
Forwarding.eca, which is where the proof part begins (and which
requires UCCore). These are now abstract theories (.eca), and so must
be cloned before being used by another theory.

In Forwarding_Code.ec, names of variable, types and constructors now
have "_" prepended to them, so that, e.g., if "self" or "adv" or "m"
or "r" is used in the UC DSL code, there won't be a conflict in the
generated code. On the other hand, the top-level interface and
functionality names - which are turned into theories and modules,
respectively, are unchanged. (EasyCrypt doesn't allow them to start
with "_".)

(Work on KeyExchange comes next.)