(1) In UC DSL, made the definition of legal identifiers more restrictive, to faciliate being able to make an EasyCrypt identifier that faithfully combines multiple UC DSL identifiers, using "__" as a separator. Consequently, UC DSL identifers must neither start nor end with "_", but also may not have occurrences of "__". (They still may not begin with uc_ or UC_, and may not coincide with the list of forbidden EasyCrypt identifiers.)
(2) To faciliate generation of EasyCrypt code for message matching, in
message matching clauses like

  *     => { ... }
  D.E.* => { ... }

the "..." must now literally be

  fail.

When generating EasyCrypt code, we will always initialize the result
variable

  _r <- None;

This means that a clause whose pattern ends in * can simply be ignored. We
now never need to use more than one such clause, which will come at the very
end.

If we didn't make this change, we would have to expand out such
clauses to work through the possible messages, one by one. Instead,
when the user doesn't want "fail.", we'll make them spell out the
message pattern explicitly. Since the instruction blow will often
involve either or both of a source port and message parameters, this
will typically be needed anyway.

(3) The above changes meant that a large number of the unit tests
needed to be modified. In particular, instead of using identifiers
like D_, we're now using D'. Also added new tests for the above
changes. (The identifier restriction tests are in the lexer tests, and
the */failure tests are in the parser tests.)