(1) When a program variable was bound in a message match clause pattern or in a match statement pattern, eclipsing the variable when used in expressions, uninitialized variable error messages were still being generated (even though it wasn't the variable that was being reference, but a bound identifier). (We missed this issue in our unit tests.)
This was unavoidable without changes to EcEnv, as program variables
(for us, not EasyCrypt) and bound pattern identifiers look the same in
environment. It's confusing when variables are rebound, though,
especially since on the lhs of assignments, they are never
eclipsed. ***So changed the typechecker so that variables *may not* be
eclipsed in message match clauses or in match statements.*** This
only broke two unit tests.

(2) Added in the ucdsl examples/sessions directory the first example
of a multi-session functionality: Forwarding.