Added choice datatypes like
type ('a, 'b, 'c, 'd) choice4 = [
  | Choice4_1 of 'a
  | Choice4_2 of 'b
  | Choice4_3 of 'c
  | Choice4_4 of 'd
].

plus the correponding EPDPs.

These exist up to arity 8, as do the EPDPs for tuples.

If you want to have an EPDP for a user-defined datatype, you can
define an isomorphism between that datatype and a choice datatype,
and then piggyback on the choice EPDP. With records, you can do this
but via tuples.