Changed the typing rules for message matching in initial states of ideal functionalities and parties of real functionalities to forbid matching of adversarial messages. The semantics of the DSL is that such messages will result in failure. But the user will not have to write any message clauses to achieve this effect.
This may require small changes in existing developments. See
uc-dsl/examples/smc-case-study for an example.

Here is the rationale for the change: Before a simulator receives its
first message from the ideal functionality, it is unable to catch
messages from the adversary that the adversary thinks it is sending to
the real functionality. Instead they will flow through the simulator
to the ideal functionality. There is no way for the ideal
functionality to respond to such messages, as they are not in its
adversarial interface. Thus if we have a real functionality with any parties
that can respond to adversarial messages in their initial states, we
won't be able to have a corresponding ideal functionality and simulator
such that UC security holds.

It seems better to catch this problem at typechecking time, instead
of during a proof (or using the limited security checking tool that
is planned).