Changed UC DSL so that:
(1) real and ideal functionalities cannot send-and-transition back to
their initial states, even from their initial states (this was already
true for simulators);

(2) a send-and-transition from the initial state of an ideal functionality
must send an adversarial message (and so not a direct message, back to
the environment).

- - -

(1) is sensible since functionalities cannot respond to messages in
their initial states (then implicitly cause failure); once they have
transitioned to non-initial states and can respond to adversarial
messages (in the case of a real functionality, only if the given party
has an adversarial interface), it makes no sense to revert to the
limited state.

(2) largely pertains to ideal functionalities *with* simulators. Once
the ideal functionality has transitioned to a non-initial state, we
thus know its simulator has been woken up, and can simulate messages
coming from the adversary (which the adversary thinks are going to the
real funnctionality). Thus there is no worry that messages from the
adversary will flow through the simulator and come to the ideal
functionality, which is expecting message from its simulator.  (When
the ideal functionality is in its initial state, this can happen, but
in this case the ideal functionality is not expecting messages from
its simulator.)

Applying (2) to an ideal functionality (like a forwarder) that has
no simulator, we have that the ideal functionality's first message
must be to the adversary. This is what all of our examples do,
but does mean that we cannot write an ideal functionality that,
say, forwards an input without involving the adversary. Note though,
that the adversary could simply be pinged with no information,
letting it either return control or send control elsewhere.