Added type for (real + ideal) worlds in ucInterpreter.ml. This will need a pretty printer. (Conversion of real functionality expression to corresponding world will come next.)
Added types for unit information in ucTypedSpec, along with function that
takes a root to its unit information. Unit information includes
the numbers of adversarial port indices needed by a unit, as well
as functions for mapping a base port index and served adversarial id to
its port index, and for mapping a base port index and subfunctionality id
to its port index.