Changed UCBasicTypes.ec so in uc-dsl/prelude there is just a link to the version in common, which exposes the structure of port, and where msg and mode are included in this theory. That way all the definitions needed by the EC code produced by the upcoming translator will be in this file. E.g., a DSL spec used msg or Adv, it would get the real versions of these definitions (which would be useless, but not dangerous), and so the generated EC code would do just the same thing. Note that there is no good reason for UC DSL specification to drill down to the structure of ports, but this is also not dangerous, and it could be done indirectly anyway via the supplied epdp from ports to univ.
Also added drafts of a lot of comments in this file, talking in particular
about how port indices are chosen.

An unrelated change: make it so in send and transition instructions,
if the destination port isn't an identifier, then parentheses are
required. This is easier on the human eye, and in line with what we
are doing with sent message expressions in the deploy-interpreter
branch.