Added new function to UC DSL addition to EcCommands, to support the interpreter's `load` command; it's the last of these functions (other comments are improved):
(* push new scope onto stack, created from the previous one, but
   reverting the environment and required theories to the ones
   of the prelude *)
val ucdsl_new : unit -> unit

(* end scope, reverting to saved one from stack, which is updated to
   include required theories of ended scope *)
val ucdsl_end : unit -> unit

(* end scope, reverting to saved one from stack, but ignoring the
   scope being ended *)
val ucdsl_end_ignore : unit -> unit

When we execute a load command, we run ucdsl_new (), and then run
parse_and_typecheck_file_or_id of UcParseAndTypecheckFile with
an argument (FOID_Id id), where id is a localized root.

This can be repeated, growing the stack of scopes. We have to keep track
manually of the stack of maps created by typechecking.

When we execute an `undo` command that takes us back to a previous `load`,
we run ucdsl_end_ignore () to simply pop the scopes stack. But we have
to manually pop the maps stack, uncovering the typechecking maps (for
interfaces, functionalities and simulators) that we are reverting to.