Changed
val send_message_to_real_or_ideal_config :
      config -> sent_msg_expr -> config * effect

to take a parse tree form of sent message expression. That way typechecking
is done as part of this function, meaning that we don't need a function
in the .mli for returning the environment of a configuration. (This
was already true for adding assumption to the global context.)

Fixed bugs in typechecking and pretty printing of sent message expressions.
We can now have, e.g., T.foo as source or destination port without needing
parentheses either in parsing or pretty printing.

In the message path of a send message expression we now only allow
terminating with a basic interface when this is the adversarial
interface of an ideal functionality. Otherwise, one must explicitly
use a composite interface and then sub-interface (whose meaning is
a basic interface).

Fixed bugs in typechecking and pretty printing of sent message expressions.
In the pretty printing