Changed comments about tags of messages in UCBasicTypes to read:
"when translating from the UC DSL, we assign tags to messages of
interfaces so that messages with different message paths are
guaranteed to be distinct"

This requires a global system for assigning tags, but we already do such
a thing for adversarial port indices.

his will bring the correspondence between the behavior of the
interpreter and the generated EasyCrypt code closer. It'll still be
the case that in the latter the environment or adversary can express
messages that can't be expressed in the DSL. E.g., the universe value
for a message with a given tag might not decode successfully. But such
messages will result in failure. On the other hand, with this change,
a DSL-expressable message can never be accepted where a message
with another message path is expected.