Saving work for debugging an SMT problem connected with interpretation.
The interpreter test test0_of_UcInterpreter.ml + the code in the
interpreter is now focused on generating a simple SMT goal that
should succeed. The correct functionality is commented out.

Note: we need to take the dump_form code out of ecCoreFol.ml/.mli
once this is sorted out. The interpreter is currently using
this function.