Attempted to redesign message tags in UCBasicTypes to be assignable unit by unit and to encode just the information relating to message paths that is not already implicit in a message. See the documentation there.
Ordinary sent message expressions are redudant in that their destination
port indices - for in messages - and their source port indices - for
out messages - should be determined from the message paths. So
send_message_to_real_or_ideal_config now checks for this consistency,
signaling an error otherwise. And internally generated messages should
be created to be consistent.

A consequence of this is that when sending a message to a functionality,
we never have to use the destination port indices (or source ones, in the
case of messages from subfunctionalities or argument to parents) for
routing - instead we can just use the message paths. This is good,
because it doesn't rely on formula evaluation.