Fixed bug in typechecker (!) pertaining to ordinary match clauses like
  A i b => { ... }

We forgot to test the case where there are multiple args (!), but it
was an easy fix. We should add a test case like this to our test suite.

The ordinary match instruction is now implemented in the interpreter, but
is disabled, awaiting a change to the type and behavior of
UcEcFormEval.deconstruct_data.