Updated top-level README.
Added simplification hint turning xs ++ [] -> xs. Will turn into a
targetted rewriting hint when we can use them in the interpretation.

Took some of the junk out of the SMC2 example, although we still
have a party pinging the adversary, as an example of how that works.
There is now a complete ideal world .uci script. The real script
blocks due to an epdp simplification failure - waiting either
for being able to use rewriting hints of the EasyCrypt bug to be fixed.

Up until the blockage point, the assertions in both .uci files are
identical.