Changed UcInterpreter so it now maintains a list of rewriting databases, to be used in simplification.
These start out being UCEncoding.epdp and
UCBasicTypes.uc_dsl_interpreter_hints.

There are functions in the UCInterpreter API allowing rewriting
databases to be removed or added or pretty-printed. Just as
step_running_or_sending_real_or_ideal_config already took an optional
value of type prover_infos, it now takes an optional pair of lists of
rewriting databases. If this is Some(rm, add), the rewriting databases
used during this step (but discarded after that, i.e., not put into
the new configuration) are the result of removing the databases rm and
then adding those in add.

The API functions in UcEcFormEval now take an extra argument: the
left-to-right rewriting lemmas that should be used repeatedly in
simplification. Nothing is currently done with these extra arguments.