We already had that ideal functionalities in their initial states that were executing a send-and-transtion instruction had to send the message to their simulator (waking it up) or the adversary (if no simulator).
This commit makes parties of real functionalities that serve basic
adversarial interfaces behave the same: if such a party is executing
a send-and-transition instruction in its initial state, it must
go to the adversary.

Because adversarial messages are rejected in the initial states
of ideal functionalities and parties of real functionalities, this
means that at the point when the adversary *could* send a message
to an ideal functionality or a party of a real functionality, it
will already have received a message from that entity.

Thus if it didn't already know the adverarial port index used
in communication with that entity, it can learn it from the
source address of the message it gets.

(Remember that for ideal functionalties, parties of real
functionalities and simulators, transitions back into initial states
are always disallowed.)