Tracked upcoming change to EasyCrypt (development branch opaque-op-smt) that gets rid of nosmt option for both lemmas and concrete operators.
For operators, there is a new tag [smt_opaque] that makes operator
bodies opaque to SMT. Using this in EasyUC EC files caused no
problems.

The change to lemmas can be worked around in EC files by controlling
arguments to smt(...).

In the interpreter scripts, it might be necessary to use wanted lemmas in
prover options in step and run commands, but the SMC2 example still works
without them.