Added a new feature to the UC DSL: ideal functionalities in singleton units may now have no adversarial interfaces, modeling that the adversary learns nothing from their execution.
For uniformity, these ideal ideal functionalities are still not
allowed to transition back to their initial states.

The requirements of ideal functionalities in triple-units are
unchanged. If a real functionality has no adversarial interfaces, even
via subfunctionalities, it still needs a simulator (which will be
woken up by its ideal functionality), but it will never send any
messages to the adversary or receive any messages from the
adversary.

When a real functionality has parameters, the simulator *will* have to
simulate the messages coming in and out of the ideal functionality
arguments as the real functionality executes.