The unit test suite (uc-dsl/testing/dsl-test-suite) for ucdsl now returns a status of 0 when all tests pass, and 1 when at least one test failed (or there was some other error).
We had to fudge one test to make it pass both on macOS and on other
Posix systems. The problem is that on macOS there isn't a way (or at
least an easy way) to find out using an OCaml library whether the
initial letter of a file is in uppercase or lowercase. Consequently an
EasyCrypt error message will not necessarily have the correct case for
a filename. The EasyCrypt code that produces the error message tries
to figure out the capitalization of a theory, e.g., Who, by first
seeing if who.ec exists, and if it doesn't, trying Who.ec. On
non-macOS Posix platforms, this will determine the correct filename,
and so the error message will have the correct filename. But on macOS
if the file is actually Who.ec, the EasyCrypt code will say that it's
who.ec, and so the error message will be slightly wrong. The fudge was
to make a file in a test be "who.ec", so the test passes on all
platforms. This is not a serious problem, so we'll just ignore it (as
does EasyCrypt).