Tracked EasyCrypt bug fix regarding searching for EC theories and making this and generated errors stable on both case-insensitive (e.g., Linux, normally) and case-insensitive / case-preserving filesystems (e.g., macOS, normally), and then applied the same solution to searching for UC files when executing uc_requires.
In the UC file case, when, e.g., we are running `uc_requires Foo`:

When searching in a directory, we first look for the uncapitalization
(just the first letter) of Foo concatenated with ".uc" (i.e.,
"foo.uc"), and if that doesn't exist we look for Foo concatenated with
".uc" (i.e., "Foo.uc").

In a case-sensitive filesystem (e.g., Linux, normally) both foo.uc and
Foo.uc could exist, in which case we will find foo.uc (and ensure that
error messages use the correct case). BUT NOTE that when a repo with
these files is cloned on a case-insensitive filesystem, one of these
files will be lost. So this configuration SHOULD BE AVOIDED.

In a case-insensitive / case-preserving filesystem (e.g., macOS,
normally), only one of foo.uc and Foo.uc can exist; we find the one
that exists, and ensure that error messages use the correct case.

Not that, e.g., "fOO.uc" will *not* be found by `uc_requires Foo` on
either kind of filesystem. (In the case-sensitive case, this is
simply because this file does not exist. In the case-insensitive case,
"fOO.uc" is opened when we open "foo.uc", but nonetheless we do not
resolve Foo to this file, instead reporting an error.)