In the pretty printing of real worlds during a real experiment, the adversarial port indices optionally associated with parties (only when they serve a basic adversarial interface) and always with subfunctionalities are now displayed, as in (for smc2-extra in the examples subdirectory):
real world:
SMC2.SMC2Real:1
[Pt1: WaitReq, Pt2: WaitFwd1, Pt3(2): WaitFwd3, Fwd1(3): Init, Fwd2(4): Init,
 Fwd3(5): Init, Fwd4(6): Init]
(SMC.SMCReal:7[Pt1: WaitReq, Pt2: WaitKE1, Fwd(8): Init]
 (KeyExchange.KEReal:9
  [Pt1: WaitReq1, Pt2: WaitFwd1, Fw1(10): Init, Fw2(11): Init]),
 SMC.SMCReal:12[Pt1: WaitReq, Pt2: WaitKE1, Fwd(13): Init]
 (KeyExchange.KEReal:14
  [Pt1: WaitReq1, Pt2: WaitFwd1, Fw1(15): Init, Fw2(16): Init]))

Note that parties Pt1 and Pt2 of SMC2Real are not annotated with
adversarial port indices, but that Pt3 is. (In fact, it's the only
party in this example at any level that is.)

Subfunctionalties have adversarial port indices allocated for them,
even when they have no adverarial interfaces - as is the case with
Fwd3 of SMC2Real (in that case, any incoming adversarial messages
will result in failure).

More annotations could be made, but would clutter things up too much.
Subaddresses are easy to calculate. And the external port indices for
direct and adversarial interfaces of parties, although calculated
using the lexicographic order of the subinterfaces served by the
parties, can be omitted using the $ (as opposed to @) notation. So
annotations for them are less necessary.