Canonical ports may not be simplified if an assumption is introduced equating func to some value, e.g.,
foo : func = [1; 2]

Then (func ++ [3], 1) should simplify to ([1; 2; 3], 1).

However, when given a message to send, the interpreter assumed
canonical ports were already simplified.

This commit fixes this bug by tracking whether func remains *abstract*,
according to a simple sufficient but not necessary condition: apart
from the automatically introduced assumption that func and adv are
incomparable, it may only appear freely in envport formulas.

When this condition is violated, the user is warned that interpetation
may become slower (as ports may not remain canonical).