Added to the interpreter a notion of being  *weakly simplified* that is strong enough for values (like arguments to messages and states) that we won't *immediately* need to make decisions about, but less expensive than invoking simplification.
We start with values made out of constructors and integer literals,
but then we also allow leaves (not lhs's of applications) that are

* identifiers in the global context, even though they might be
  rewritten by assumptions during substitution

* of the form (func ++ <int list literal>), but just when func is
  abstract (see previous commit)

* the constants env_root_port, adv and adv_root_port

When running send and transitions instructions, we settle for
arguments to messages and states being only weakly simplified. When
they are used later in boolean expressions of conditionals or
expressions to be matched or source of destination ports, they will be
fully simplified.