(1) Added completed composition theorem theory (common/UCComposition.ec).
(2) Tracked changes to EasyCrypt, including its steps to unify
expressions and formulas. In the UC DSL, we're taking a different
approach, because we don't need formulas relating to program judgments
and related things.  Consequently, we are only parsing the expressions
we need, and doing our own typechecking. We continue to control our
own concrete syntax, joining the EasyCrypt types
post-typechecking. (If EasyCrypt goes on to drop the type expr
(abstract expressions), we'll add it back in in UcTypedSpec, and
provide injections and projections to/from form.) This commit adds
documentation relating to the above.

Also dropped some debugging start symbols to the parser, and dropped
some debugging executables.