EasyCrypt has made a move to getting rid of `expr` as a type, instead using `form`. Continuing changes to EasyCrypt made it clear that we need to track this change, and so this commit does this.
The commit makes minimal changes to our grammar, so the "expressions"
our DSL uses are unchanged, syntactically. But then I changed the
parser to jump directly to `pformula`, and used `trans_form_opt` for
typechecking the subset of `pformula` that can be expressed. We
now don't use `expr` at all, even though it - so far - still exists
as a type.

Small changes in the UC DSL typechecker and interpreter were needed,
e.g., using f_subst instead of e_subst. But it was pretty straightforward.
Some error messages have changed, as we are now using standard EasyCrypt
code for these messages.