When processing -I options, we now trim trailing /s. Not doing this was triggering a bug in edge cases like
ucdsl -I A/ -I B A/Foo.uc

For this, we want the directory A to have the highest precedence for
both uc_require and ec_require, because of the directory of the .uc
file. But the presence of the / in `-I A/` was stopping this from
working, as A/ and A weren't seen to be the same.