Changed the string for an internal port from "intport:Party" or "intport:RealFun.Party" to "intport Party" or "intport RealFun.Party".  The parser only produces such strings as arguments to PFident when given intports - otherwise they never have spaces in them.
This was needed because we are now using the standard EasyCrypt
typechecker and error reporter for formulas, and otherwise the
error messages looked different from the concrete syntax, as
they involved colons, not spaces.

Now we get, e.g.,

  unknown variable or constant: `intport Q'

This is the best we can do without changing the error reporting.