Wrote a module implementing EasyCrypt "comment macros", macros defined as special EasyCrypt comments.
The module can extract the macros embedded as comments in an EasyCrypt
file, and then apply them to arguments, substituting those arguments
for the macro's parameters in its body.

E.g., one could define this macro in connection with the security
bound of an IND-CPA security scheme:

(*! SecBound(Adv, Mem)
`|Pr[GRF(PRF, Adv2RFA(<<Adv>>)).main() @ <<Mem>> : res] -
  Pr[GRF(TRF, Adv2RFA(<<Adv>>)).main() @ <<Mem>> : res]| +
(limit_pre%r + limit_post%r) / (2 ^ text_len)%r
*)

Then (using a function of the module) one can apply SecBound to
an adversary and memory.

An argument to a macro could be used to supply an EasyCrypt path Path,
and then part of a body could be, e.g., <<Path>>.PRF, <<Path>>.TRF,
etc., if PRF, TRF, etc. were not defined appropriately in
the scope in which the application of SecBound is to be used.