Finished revision of dummy adversary and simulator wrapper, along with the proof of the dummy adversary theorem. Also improved some uses of the transitivity tactic in the composition theorem proof. And added two new lemmas to UCListPO.
The loop of the simulator now is:

  while (not_done) {
    if (m.`2.`1 = adv /\ m.`2.`2 <> sim_adv_pi) {
      r <@ Adv.invoke(m);
      (r, m, not_done) <@ after_adv(r);
    } else {
      (* if_addr_opt <> None *)
      r <@ Core.invoke(m);
      (r, m, not_done) <@ after_core(r);
    }
  }

(before there were two branchings that led to calls to Core.invoke).

Changed invoke lemma for ideal functionalities to look like:

 forall (n : int),
 hoare
 [IdealFunc.invoke :
  invar_if (glob IdealFunc) /\ term_if (glob IdealFunc) = n ==>
  invar_if (glob IdealFunc) /\
  (res <> None =>
   term_if (glob IdealFunc) < n /\
   ((oget res).`1 = Adv =>
    (oget res).`2.`2 = sim_adv_pi /\ (oget res).`3.`2 = 1))]

so we are also insisting that the source port index of an outgoing
adversarial message is 1. (This helps in the proof of the dummy
adversary theorem.)

The invoke lemma for the core of the simulator doesn't track port
indices at all.

The old MakeSimulator theory is now changed to Simulator, as
it includes the combined environment and the proof of
the dummy adversary theorem.